Changes to the XHTML output.
It now also works with lists of dictionaries. It still works with the
previously returned proxy results. In the case of a list of
dictionaries, the order of the keys, will represent the table headers,
will be alphabetically.

Also changes the normalize_whitespace() helper function to a generic xml
whitespace normalization function.

Added two short XHTML output tests.