Refactored the REST resource type folder.
Hopefully the flow of HTTP requests is clearer now.

Also changed the handling of GET request for the actual resource folder.
It now returns a list with the indexed values of all resources in the
folder. The old behaviour was to return a list with all registered
indexes and resources. This is now available by requesting the ".meta"
resource of any resource folder. I also had to change some tests so they
use the ".meta" resource.